requirements.txt
Django>=3.1,<4.0
djangorestframework>=3.12.0,<4.0
django-allauth>=0.42.0
dj-rest-auth
psycopg2>=2.8.6
requests
